+++ packages/design-system-web/src/components/geov-yasgui/geov-yasgui.tsx
@@ -4,13 +4,15 @@ import { importYasgui } from '../../lib/importYasgui';
 import generatePluginMapCircles, { MapCircleConfig } from './PluginMapCircles';
 import { closeOutline, settingsOutline } from 'ionicons/icons';
 import { isNode } from '../../lib/isNode';
+import generatePluginTimeline, { TimelineConfig } from './PluginTimeline';
 
-export type CustomPlugin = 'mapCircles';
+export type CustomPlugin = 'mapCircles' | 'timeline';
 export type BuiltInPlugin = 'response' | 'table';
 export type Plugin = BuiltInPlugin | CustomPlugin;
 
 export interface PluginConfig {
   mapCircles?: MapCircleConfig;
+  timeline?: TimelineConfig;
 }
 
 export interface QueryTab {
@@ -149,6 +151,7 @@ export class GeovYasgui {
   registerCustomPlugins() {
     const customPlugins: { [key in CustomPlugin]: any } = {
       mapCircles: generatePluginMapCircles(this.pluginConfig.mapCircles),
+      timeline: generatePluginTimeline(this.pluginConfig.timeline),
     };
 
     this.plugins?.forEach(plugin => {